requirements.txt
streamlit>=1.38.0
pydantic>=2.8.0
python-dotenv>=1.0.1
httpx>=0.27.0
openai>=1.42.0
huggingface_hub>=0.24.6

